implement facade pattern
