+++ src/core-api.client.ts
@@ -1,19 +1,19 @@
-import { fetch, toBase64, URLSearchParams } from './polyfills';
 import { DTOName, DTOModels } from './core/dto-name.model';
 import { ClientConfig } from './core/client-config.model';
 import { OauthTokenResponse } from './core/oauth-token-response.model';
 import { ClientResponse } from './core/client-response.model';
 import { BatchAction } from './core/batch-action.model';
 import { RequestOptionsFacory } from './core/request-options.facory';
-import { BatchRequest } from './core/batch-request.model';
-import { BatchResponse, BatchResponseJson } from './core/batch-response';
-import { ErrorResponse, HttpRequestOptions, HttpResponse } from './core/error-response.model';
+import { BatchResponseJson } from './core/batch-response';
+
+import { AuthService } from './core/oauth-token.service';
+import { HttpService } from './core/http-service';
+import { ClientService } from './core/client.service';
 
 export class CoreAPIClient {
 
-  private _token: OauthTokenResponse | undefined;
-
-  private _config: ClientConfig = {
+  private _client: ClientService;
+  private _config_default: ClientConfig = {
     debug: false,
 
     oauthEndpoint: 'https://ds.coresuite.com/api/oauth2/v1',
@@ -54,125 +54,12 @@ export class CoreAPIClient {
    * @returns CoreAPIClient
    */
   constructor(config: ClientConfig) {
-    this._config = Object.assign(this._config, config);
+    const _config = Object.assign(this._config_default, config);
+    const _http = new HttpService(_config);
+    const _auth = new AuthService(_http, _config);
+    this._client = new ClientService(_config, _http, _auth);
   }
 
-  private async _fetchAndSaveToken() {
-    const body = new URLSearchParams({
-      grant_type: this._config.authGrantType,
-      ...(this._config.authGrantType === 'password'
-        ? {
-          username: `${this._config.authAccountName}/${this._config.authUserName}`,
-          password: this._config.authPassword
-        }
-        : {}
-      )
-    });
-
-    const response = await this._request<string>(`${this._config.oauthEndpoint}/token`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'Accept': 'application/json',
-        'Authorization': `Basic ${toBase64(`${this._config.clientIdentifier}:${this._config.clientSecret}`)}`,
-        ...RequestOptionsFacory.getRequestXHeaders(this._config)
-      },
-      body: body.toString()
-    });
-
-    if (this._config.debug && this._config.tokenCacheFilePath) {
-      try {
-        const fs = require('fs'); // inline import for isomorphic
-        fs.writeFileSync(this._config.tokenCacheFilePath, JSON.stringify(response));
-      } catch (error) {
-        console.error(`ERROR: could not create ${this._config.tokenCacheFilePath}`, error);
-      }
-    }
-
-    return typeof response === 'string' ? JSON.parse(response) : response;
-  }
-
-  private async _readToken(): Promise<OauthTokenResponse> {
-    try {
-      return await new Promise<OauthTokenResponse>((resolve, fail) => {
-        if (this._config.debug && this._config.tokenCacheFilePath) {
-          const path = require('path');
-          return resolve(require(path.resolve(this._config.tokenCacheFilePath)));
-        }
-        fail({ code: 'MODULE_NOT_FOUND' });
-      });
-    } catch (error) {
-      if (error.code === 'MODULE_NOT_FOUND') {
-        return await this._fetchAndSaveToken();
-      }
-      throw error;
-    }
-  }
-
-  private _ensureToken(): Promise<OauthTokenResponse> {
-    return this._token
-      ? Promise.resolve(this._token)
-      : this._readToken()
-        .then(token => this.setToken(token).getToken() as OauthTokenResponse);
-  }
-
-  private _request<T>(uri: string, options: HttpRequestOptions): Promise<T | string | null> {
-    if (this._config.debug) {
-      console.log(`[httpRequest] outgoing ${uri} options[${JSON.stringify(options, null, 2)}]`);
-    }
-    return fetch(uri, options)
-      .then(async (response: HttpResponse) => {
-
-        const contentType = response.headers.get('content-type');
-        const isJson = contentType && contentType.includes('application/json');
-
-        const content: T | string | null = await (!!response.json && !!response.text
-          ? isJson
-            ? response.json<T>()
-            : response.text()
-          : Promise.resolve(null)
-        );
-
-        if (!response.ok && [304, 302].indexOf(response.status || -1) === -1) {
-          throw <ErrorResponse<any, HttpRequestOptions>>{
-            statusCode: response.status,
-            message: response.statusText,
-            error: content,
-            response: response,
-            options: options
-          };
-        }
-
-        if (this._config.debug) {
-          console.log(`[httpRequest] incoming going options[${JSON.stringify(options, null, 2)}] response[${JSON.stringify(content, null, 2)}]`);
-        }
-
-        return content as T;
-      });
-  }
-
-  private async _requestDataApi<T extends DTOModels>(
-    method: 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE',
-    dtoName: DTOName,
-    dtoData: T | null | any,
-    dtoId: string | null = null,
-    additionalQs: { [k: string]: string | number | boolean } = {}
-  ) {
-    const token = await this._ensureToken();
-
-    const params = new URLSearchParams(Object.assign({},
-      RequestOptionsFacory.getRequestAccountQueryParams(token, this._config),
-      method !== 'DELETE' ? { dtos: RequestOptionsFacory.getDTOVersionsString([dtoName]) } : undefined,
-      additionalQs) as { [key: string]: string });
-
-    return await this._request<T>(
-      `${RequestOptionsFacory.getDataApiUriFor(token, dtoName, dtoId)}?${params}`,
-      {
-        method,
-        headers: RequestOptionsFacory.getRequestHeaders(token, this._config),
-        body: method != 'GET' ? JSON.stringify(dtoData) : undefined
-      });
-  }
 
   /**
    * Creates UUID
@@ -189,20 +76,7 @@ export class CoreAPIClient {
    * @returns Promise<{ data: DTO[] }>
    */
   public async query<T extends { [projection: string]: DTOModels }>(coreSQL: string, dtoNames: DTOName[]): Promise<{ data: T[] }> {
-    const token = await this._ensureToken();
-    const params = new URLSearchParams({
-      ...RequestOptionsFacory.getRequestAccountQueryParams(token, this._config),
-      dtos: RequestOptionsFacory.getDTOVersionsString(dtoNames)
-    });
-    return await this._request(`${token.cluster_url}/api/query/v1?${params}`,
-      {
-        method: 'POST',
-        headers: Object.assign(
-          { 'Content-Type': 'application/json' },
-          RequestOptionsFacory.getRequestHeaders(token, this._config)
-        ),
-        body: JSON.stringify({ query: coreSQL })
-      }) as { data: T[] };
+    return this._client.query<T>(coreSQL, dtoNames);
   }
 
   /**
@@ -212,13 +86,11 @@ export class CoreAPIClient {
    * @returns Promise<ClientResponse<DTO>>
    */
   public async getById<T extends DTOModels>(resourceName: DTOName, resourceId: string): Promise<ClientResponse<T>> {
-    const response = await this._requestDataApi('GET', resourceName, null, resourceId);
-    return typeof response === 'string' ? JSON.parse(response) : response; // not sending json headers
+    return this._client.getById<T>(resourceName, resourceId);
   }
 
   public async deleteById<T extends Partial<DTOModels>>(resourceName: DTOName, resource: { id: string, lastChanged: number }): Promise<undefined> {
-    const { id, lastChanged } = resource;
-    return this._requestDataApi('DELETE', resourceName, null, id, { lastChanged }) as any as Promise<undefined>;
+    return this._client.deleteById<T>(resourceName, resource);
   }
 
   /**
@@ -228,7 +100,7 @@ export class CoreAPIClient {
    * @returns Promise<ClientResponse<DTO>>
    */
   public async post<T extends DTOModels>(resourceName: DTOName, resource: T): Promise<ClientResponse<T>> {
-    return this._requestDataApi('POST', resourceName, resource) as Promise<ClientResponse<T>>;
+    return this._client.post<T>(resourceName, resource);
   }
 
   /**
@@ -238,8 +110,7 @@ export class CoreAPIClient {
    * @returns Promise<ClientResponse<DTO>>
    */
   public async put<T extends DTOModels>(resourceName: DTOName, resource: T): Promise<ClientResponse<T>> {
-    const { id } = resource;
-    return this._requestDataApi('PUT', resourceName, resource, id) as Promise<ClientResponse<T>>;
+    return this._client.put<T>(resourceName, resource);
   }
 
   /**
@@ -250,8 +121,7 @@ export class CoreAPIClient {
    * @returns Promise<ClientResponse<DTO>>
    */
   public async patch<T extends DTOModels>(resourceName: DTOName, resource: T): Promise<ClientResponse<T>> {
-    const { id } = resource;
-    return this._requestDataApi('PATCH', resourceName, resource, id) as Promise<ClientResponse<T>>;
+    return this._client.patch<T>(resourceName, resource);
   }
 
 
@@ -271,27 +141,7 @@ export class CoreAPIClient {
    * @returns Promise<BatchResponseJson<T>[]>
    */
   public async batch<T extends DTOModels>(actions: BatchAction[]): Promise<BatchResponseJson<T>[]> {
-    const token = await this._ensureToken();
-    const body = new BatchRequest(token, this._config, actions).toString();
-
-    const params = new URLSearchParams(Object.assign({
-      clientIdentifier: this._config.clientIdentifier,
-      dtos: RequestOptionsFacory.getDTOVersionsString(actions.map(it => it.dtoName)),
-      ...RequestOptionsFacory.getRequestAccountQueryParams(token, this._config),
-    }));
-
-    const responseBody = await this._request<string>(`${token.cluster_url}/api/data/batch/v1?${params}`, {
-      method: 'POST',
-      headers: {
-        'content-type': 'multipart/mixed;boundary="======boundary======"',
-        ...RequestOptionsFacory.getRequestHeaders(token, this._config),
-      },
-      body
-    });
-
-    return responseBody !== null
-      ? new BatchResponse<T>(responseBody).toJson()
-      : [];
+    return this._client.batch<T>(actions);
   }
 
 
@@ -300,7 +150,7 @@ export class CoreAPIClient {
    * @returns Readonly<OauthTokenResponse> | undefined
    */
   public getToken(): Readonly<OauthTokenResponse> | undefined {
-    return this._token;
+    return this._client.getToken();
   }
 
   /**
@@ -309,12 +159,7 @@ export class CoreAPIClient {
    * @returns CoreAPIClient
    */
   public setToken(token: OauthTokenResponse): CoreAPIClient {
-
-    if (!token || !token.account) {
-      throw new Error('invalid token');
-    }
-
-    this._token = token;
+    this._client.setToken(token);
     return this;
   }
